changed circular trail against a flexible trail

+++ src/de/javavorlesung/pong/Coordinate.java
@@ -13,16 +13,16 @@ public class Coordinate {
 
 
 
-    public double getX() {
-        return x;
+    public int getX() {
+        return (int)x;
     }
 
     public void setX(double x) {
         this.x = x;
     }
 
-    public double getY() {
-        return y;
+    public int getY() {
+        return (int)y;
     }
 
     public void setY(double y) {
